--- students/sec1814080911106/.idea/.name
@@ -1,1 +1,0 @@
-Sec1814080911106Activity
