--- students/sec1814080911106/app/.gitignore
@@ -1,1 +1,0 @@
-/build
