--- students/sec1814080911106/app/src/main/res/values/strings.xml
@@ -1,3 +1,0 @@
-<resources>
-    <string name="app_name">图片浏览</string>
-</resources>
