--- students/sec1814080911106/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Sun Oct 18 15:34:22 CST 2020
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-6.1.1-all.zip
