--- students/sec1814080911106/settings.gradle
@@ -1,2 +1,0 @@
-include ':app'
-rootProject.name = "Sec1814080911106Activity"
